Turbulence visualization
Started with the identification of coherent structures by visualizing the Q-criterion isosurfaces over the bottom half of the domain in ParaView. Regarding the data post-processing with MATLAB, further turbulent statistics, such as the velocity fluctuations and the mean Kinetic Turbulent Energy, have been computed. Proper documentation of the code is still yet to be implemented.